Use deterministic unique ids in Descriptions
Rerunning failed tests with Surefire requires a method to identify the tests.
Surefire currently uses a tuple of (Class, Method) to identify failed tests.
Cucumber-jvm uses test suites which do not consist of Classes or methods.
Therefore another method is needed.

JUnit provides the option to select tests that match a Description.
Descriptions are compared using an unique Id. To identify tests between
different runs we should provide Descriptions with a predictable unique
id.

The current implementation did not suffice. While the provided cucumber
elements are unique, they are not predictable. Each run would create a new
instance of the identifier making it impossible to use their descriptions
to rerun a failed test.

After this change it will be possible to update Surefire to use
descriptions to rerun failed tests.

  Related Issues:
  - https://issues.apache.org/jira/browse/SUREFIRE-1372
  - https://github.com/cucumber/cucumber-jvm/issues/1120
  - https://github.com/temyers/cucumber-jvm-parallel-plugin/issues/31